requirements.txt




google-cloud-aiplatform
streamlit
google.cloud.logging